+++ R/get_dates.R
@@ -1,6 +1,6 @@
 #' Retrieve dates component of a BirdFlow model
 #'
-#' `get_dates()` retrieves the dates component of a BirdFlow model.
+#' `get_dates()` get date information for a BirdFlow model
 #'
 #'  \pkg{BirdFlowR} uses the same conventions for assigning dates to
 #'  eBird "weeks" as \pkg{ebirdst}. This changed with the 2022 release.
@@ -25,12 +25,30 @@
 #'
 #' [truncate_birdflow()] creates models with a subset of the intervals
 #' and renames them `1:[n_timesteps()]` of the truncated model so if it
-#' has been used the data may differ from `ebirdst::ebirdst_weeks`.
+#' has been used `week` and `timestep` will not have identical values.
+#'
+#'
 #'
 #' @param bf A BirdFlow object to retrieve dates from.
 #'
+#' @return It returns a data frame.
 #'
-#' @return A data frame with:
+#' The columns are:
+#' \item{timestep}{The model timestep associated with each row. Will always
+#' equal the row number.}
+#' \item{date}{The date associated with the midpoint of the timestep}
+#' \item{label}{ The month and day based label associated with the interval.
+#' Consistent with eBird's 2022 date scheme will be one day off the `date`
+#' after February on leap years but are still used to label intervals. If the
+#' model was fit with an older version of \pkg{ebirdst} they will never be
+#' offset}
+#' \item{julian}{The Julian date (day of year) associated with the timestep
+#' center}
+#' \item{week}{The eBirdst week number associated with the date. For full year
+#' models this is identical to `timestep` but after
+#' [truncation](truncate_birdflow) they will differ.
+#'
+#' Prior to \pkg{BirdFlowR} v. 0.1.0.9040 it returned columns:
 #' \item{interval}{The interval or timestep associated with each date.
 #' It will range from 1 to [n_timesteps()].  With full models this is
 #' equivalent to the week of the year but with truncated models may
@@ -41,5 +59,22 @@
 #' \item{doy}{The day of year associated with the midpoint of each interval.}
 #' @export
 get_dates <- function(bf) {
- return(bf$dates)
+
+  # Return models fit with ebirdst 3.2022.0 and newer as is
+  if (get_metadata(bf, "ebird_version_year") >= 2022)
+    return(bf$dates)
+
+  # Reformat older dates (ebirdst 2021)
+  dates <- bf$dates
+  names(dates)[names(dates) == "interval"] <- "timestep"
+  if (!"week" %in% names(dates) && nrow(dates) == 52)
+    dates$week <- 1:52
+  d <- lubridate::as_date(dates$date)
+  dates$label <- paste(
+    lubridate::month(d, abbr = FALSE, label = TRUE),
+    lubridate::mday(d), sep = " ")
+  dates$julian <- lubridate::yday(d)
+  dates <- dates[, c("timestep", "date", "label", "julian", "week")]
+  return(dates)
+
 }