+++ R/truncate_birdflow.R
@@ -27,11 +27,18 @@
 truncate_birdflow <- function(bf, ...) {
 
   # Add timestep_padding metadata if it doesn't exist
-  # (back compatability code)
+  ### back compatability
   if (is.null(bf$metadata$timestep_padding)) {
     ts_padding <- get_timestep_padding(bf)
     bf$metadata$timestep_padding <- ts_padding
   }
+
+  # Add dates$weeks if it doesn't exist
+  ### back compatibility code
+  if(is.null(bf$dates$week) && nrow(bf$dates) == 52){
+    bf$dates$week <- 1:52
+  }
+
 
   # Handle special case of circular preprocessed model
   # The last date and distribution in these models is a duplicate of the